feat(referentiels): raccorde la modale de confirmation et ré-organise le bandeau lié à la bascule
