+++ apps/app/src/referentiels/switch-to-te.banner/switch-to-te-info.button.tsx
