+++ apps/app/src/referentiels/switch-to-te.banner/switch-to-te-unauthorized.banner.tsx
@@ -1,5 +1,5 @@
 import { appLabels } from '@/app/labels/catalog';
-import { ModeBannerAlert } from './mode-banner-alert';
+import { ModeBannerAlert } from '../referentiel-mode/mode-banner-alert';
 import { SwitchToTeInfoButton } from './switch-to-te-info.button';
 
 export const SwitchToTeUnauthorizedBanner = () => (
