+++ apps/app/src/referentiels/switch-to-te.banner/use-switch-to-te-status.ts
